property reassignment for 3rd party libraries
It's currently not possible modify an object property in a function. This makes it impossible to work with Leaflet objects and other 3rd party libraries. The list of exceptions by airbnb is focused on nodejs. In the frontend we have way too many exceptions to list and maintain in the lint rules, we would have to list and maintain every property of a leaflet instance and other libraries here.

The airbnb example doesn't even make sense as its not producing the same result.

// bad
function f1(obj) {
  obj.key = 1;
}

// good
function f2(obj) {
  const key = Object.prototype.hasOwnProperty.call(obj, 'key') ? obj.key : 1; // hello this is something entirely different ???
}

Additionally, this rule is easily bypassed by using _.set
And also this is totally legit though it reassigns the original object because we dont clone it.

// why is this ok ?
function f1(original) {
  const copy = original // just a reference
  copy.key = 1; // this should throw a lint error but it fails to track it
}